+++ modules/lmscan-frontend/src/main/scala/io/leisuremeta/chain/lmscan/frontend/controllers/msg/onTxMsg.scala
@@ -9,11 +9,15 @@ import tyrian.http.*
 
 object UnderTxMsg:
   private val onResponse: Response => Msg = response =>
+    Log.log("json")
     val json = Log.log(response.body)
 
     val parsed = parse(json) match
       case Right(r) => Right(r)
       case Left(l)  => Left(l.message)
+
+    Log.log("parsed")
+    Log.log(parsed)
 
     val deserialised =
       parsed.flatMap { json =>
